Made a file to list all things studio ghibli.
These shouldn't necessarily be cards themselves, but it might serve as nice reference material for future cards. It's structured in such a way that writing a regex to match certain elements shouldn't be too difficult for creating cards on the fly.

Note: this is currently unfinished. I plan to finish it in a future commit.